Fix & update script for 2023
Make the script compatible with the 2023 table, and sprinkle a few
little fixes on top while we're at it (these would ideally go into a
separate commit, but nah).

The most significant change here is the downloading of the HTML table in
the Makefile, instead of us having to remember the manual process.

Also, use "rolling" .ics files instead of per-year, which means people
won't have to import new calendars every year.